ログアウトのロジックを切り離した
