+++ components/Header.tsx
@@ -1,6 +1,7 @@
+import React, { FC } from "react";
 import styles from "styles/components/Header.module.scss";
 
-export default function Header() {
+const Header: FC = () => {
 	return (
 		<div>
 			<div className="navbar navbar-light bg-info">
@@ -9,4 +10,5 @@ export default function Header() {
 			<div className={styles.test}>Header.module.scssの確認</div>
 		</div>
 	);
-}
+};
+export default Header;
