+++ pages/index.tsx
@@ -1,17 +1,9 @@
 import Head from "next/head";
 import Image from "next/image";
-import Header from "../components/Header";
+import Header from "components/header";
+import Logout from "components/logout";
 import { useAuthentication } from "hooks/authentication";
-import firebase from "firebase/app";
 import Link from "next/link";
-
-const logout = async () => {
-	try {
-		await firebase.auth().signOut();
-	} catch (error) {
-		console.log(error.message);
-	}
-};
 
 export default function Home() {
 	const { user } = useAuthentication();
@@ -21,9 +13,7 @@ export default function Home() {
 			<div>投稿一覧ページ</div>
 			<div>
 				{user ? (
-					<button onClick={logout} className="btn btn-danger">
-						Logout
-					</button>
+					<Logout />
 				) : (
 					<Link href="/login">
 						<button>ログインする</button>
